pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "buddhist-rag"
version = "0.1.0"
description = "Buddhist RAG system for Taishō Canon (CBETA T/ folder) using TDD and clean architecture"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "lxml>=4.9.0",              # TEI P5 XML parsing
    "beautifulsoup4>=4.12.0",   # Alternative XML parsing
    "openai>=1.0.0",            # OpenAI embeddings
    "anthropic>=0.18.0",        # Claude API
    "chromadb>=0.4.0",          # Vector database
    "fastapi>=0.109.0",         # API framework
    "uvicorn>=0.27.0",          # ASGI server
    "pydantic>=2.5.0",          # Data validation
    "python-dotenv>=1.0.0",     # Environment variables
    "numpy>=1.24.0",            # Numerical operations
    "sentence-transformers>=2.3.0",  # Local embeddings
    "httpx>=0.26.0",            # HTTP client for testing
    "click>=8.1.0",             # CLI framework
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "ipython>=8.18.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow tests that make external API calls",
]

[tool.black]
line-length = 100
target-version = ["py310"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = ["E501"]  # Line too long (handled by black)

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["chromadb.*", "sentence_transformers.*"]
ignore_missing_imports = true
